+++ tests/src/Kernel/AuditChainMetricsTest.php
@@ -26,6 +26,8 @@ use PHPUnit\Framework\Attributes\RunTestsInSeparateProcesses;
 #[RunTestsInSeparateProcesses]
 final class AuditChainMetricsTest extends KernelTestBase {
 
+  use AuditChainSchemaTrait;
+
   /**
    * {@inheritdoc}
    */
@@ -48,7 +50,7 @@ final class AuditChainMetricsTest extends KernelTestBase {
    */
   protected function setUp(): void {
     parent::setUp();
-    $this->installSchema('audit_chain', ['audit_chain_log']);
+    $this->installAuditChainTables();
     $this->installConfig(['audit_chain']);
     $this->metrics = $this->container->get('audit_chain.metrics');
   }